% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradpsi_sp.R
\name{dpsi_dx}
\alias{dpsi_dx}
\title{Computed dpsi/dx
The inputs `x`, `beta`, `a`, `b`, and `mu` should be of the same length
This function is used for finding optimal (argmax) `x` given beta
This function is intended to be used inside `mvnrnd_wrap`}
\usage{
dpsi_dx(x, beta, veccCondMeanVarObj, NN, a, b, mu)
}
\description{
Computed dpsi/dx
The inputs `x`, `beta`, `a`, `b`, and `mu` should be of the same length
This function is used for finding optimal (argmax) `x` given beta
This function is intended to be used inside `mvnrnd_wrap`
}
