% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{FIC_reorder_maxmin}
\alias{FIC_reorder_maxmin}
\title{Univariate ordering under FIC approximation, first m chosen by maxmin ordering}
\usage{
FIC_reorder_maxmin(
  a,
  b,
  m,
  locs = NULL,
  covName = NULL,
  covParms = NULL,
  covMat = NULL
)
}
\description{
Univariate ordering under FIC approximation, first m chosen by maxmin ordering
}
\examples{
n1 <- 5
n2 <- 5
n <- n1 * n2
m <- 5
locs <- as.matrix(expand.grid((1:n1) / n1, (1:n2) / n2))
covparms <- c(2, 0.1, 0)
cov_name <- "matern15_isotropic"
a <- rep(-Inf, n)
b <- seq(from = -3, to = 3, length.out = n)
cat("The output order should be roughly increasing after", m, "numbers \n")
cat(FIC_reorder_maxmin(a, b, m, locs, cov_name, covparms))
}
