% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Vecc_cov_mat}
\alias{get_Vecc_cov_mat}
\title{Compute the covariance matrix under the Vecchia approximation}
\usage{
get_Vecc_cov_mat(covMat, m)
}
\arguments{
\item{covMat}{the original covariance matrix}

\item{m}{the number of nearest neighbors}
}
\value{
an n X n matrix
}
\description{
Compute the covariance matrix under the Vecchia approximation
}
