% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradpsi_sp.R
\name{grad_jacprod_jacsolv_idea5}
\alias{grad_jacprod_jacsolv_idea5}
\title{Gradient, Newton step, and Hessian multiply gradient of the psi function in
Idea V, computed using sparse A. For this function, `xAndBeta` should be a
vector of length 2 * n and all the returned vectors (matrices) are of
dimension 2 * n. In other words, beta_n and x_n are taken into
consideration. Meanwhile, based on Idea 5, it is obvious that beta_n = 0
and we don't need to know the value of x_n anyways}
\usage{
grad_jacprod_jacsolv_idea5(
  xAndBeta,
  veccCondMeanVarObj,
  a,
  b,
  retJac = F,
  retProd = T,
  retSolv = T,
  VAdj = T,
  verbose = T,
  mu = rep(0, length(a))
)
}
\description{
Gradient, Newton step, and Hessian multiply gradient of the psi function in
Idea V, computed using sparse A. For this function, `xAndBeta` should be a
vector of length 2 * n and all the returned vectors (matrices) are of
dimension 2 * n. In other words, beta_n and x_n are taken into
consideration. Meanwhile, based on Idea 5, it is obvious that beta_n = 0
and we don't need to know the value of x_n anyways
}
