% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradpsi.R
\name{jac_idea5}
\alias{jac_idea5}
\title{Jacobian of the psi function in Idea V}
\usage{
jac_idea5(xAndBeta, veccCondMeanVarObj, a, b)
}
\arguments{
\item{xAndBeta}{a vector whose first (n - 1) coeffs are x and second (n - 1)
coeffs are beta. (n - 1) instead of n because beta_n = 0 and x_n does
not matter. n is the dim of the MVN prob}

\item{veccCondMeanVarObj}{contains information of the conditional mean
coefficient, the conditional variance, and the NN array of the Vecchia
approximation}

\item{a}{lower bound vector for TMVN}

\item{b}{upper bound vector for TMVN}
}
\value{
a square matrix of dim 2n - 2
}
\description{
Jacobian of the psi function in Idea V
}
