% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradpsi_sp.R
\name{psi_wrapper}
\alias{psi_wrapper}
\title{Wrapper around the C function psi so that the first argument is `x`
The inputs `x`, `beta`, `a`, `b`, and `mu` should be of the same length
This function is used for finding optimal (argmax) `x` given beta
This function is intended to be used inside `mvnrnd_wrap`}
\usage{
psi_wrapper(x, beta, veccCondMeanVarObj, NN, a, b, mu)
}
\description{
Wrapper around the C function psi so that the first argument is `x`
The inputs `x`, `beta`, `a`, `b`, and `mu` should be of the same length
This function is used for finding optimal (argmax) `x` given beta
This function is intended to be used inside `mvnrnd_wrap`
}
