% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecc_cond_mean_var.R
\name{vecc_cond_mean_var}
\alias{vecc_cond_mean_var}
\title{Compute the conditional mean multiplier and the conditional variance
  under the Vecchia approximation.}
\usage{
vecc_cond_mean_var(covMat, NNarray)
}
\description{
Compute the conditional mean multiplier and the conditional variance
  under the Vecchia approximation.
}
