% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecc_cond_mean_var.R
\name{vecc_cond_mean_var_sp}
\alias{vecc_cond_mean_var_sp}
\title{Compute the conditional mean multiplier and the conditional variance
  under the Vecchia approximation. Different from `vecc_cond_mean_var`, the
  returned A is a sparse matrix.}
\usage{
vecc_cond_mean_var_sp(
  NNarray,
  covMat = NULL,
  locs = NULL,
  covName = NULL,
  covParms = NULL
)
}
\description{
Compute the conditional mean multiplier and the conditional variance
  under the Vecchia approximation. Different from `vecc_cond_mean_var`, the
  returned A is a sparse matrix.
}
