% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{Vecc_reorder}
\alias{Vecc_reorder}
\title{Univariate ordering under Vecchia approximation}
\usage{
Vecc_reorder(
  a,
  b,
  m,
  locs = NULL,
  covName = NULL,
  covParms = NULL,
  covMat = NULL
)
}
\description{
Univariate ordering under Vecchia approximation
}
\examples{
library(lhs)
library(GpGp)
library(VeccTMVN)
set.seed(123)
n <- 100
m <- 5
locs <- lhs::geneticLHS(n, 2)
covparms <- c(1, 0.1, 0)
cov_name <- "matern15_isotropic"
cov_mat <- get(cov_name)(covparms, locs)
a <- rep(-Inf, n)
b <- runif(n)
odr_TN <- TruncatedNormal::cholperm(cov_mat, a, b)$perm
rslt <- Vecc_reorder(a, b, m,
                     locs = locs, covName = cov_name,
                     covParms = covparms
)
# compare order
cat(rslt$order, "\n")
cat(odr_TN, "\n")
# check NN array
locs_odr <- locs[rslt$order, , drop = F]
cov_mat_odr <- get(cov_name)(covparms, locs_odr)
NN <- GpGp::find_ordered_nn(locs_odr, m)
cat("nn difference is", sum(rslt$nn - NN, na.rm = T), "\n")
# check A matrix and conditional var
rslt_check <- vecc_cond_mean_var_sp(rslt$nn, covMat = cov_mat_odr)
cat("A difference is", sum(rslt$A - rslt_check$A), "\n")
cat(
  "Conditional variance difference is",
  sum(rslt$cond_var - rslt_check$cond_var), "\n"
)

}
