% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VeccTMVN-package.R
\docType{package}
\name{VeccTMVN-package}
\alias{VeccTMVN}
\alias{VeccTMVN-package}
\title{VeccTMVN: Multivariate Normal Probabilities using Vecchia Approximation}
\description{
Under a different representation of the multivariate normal (MVN) probability, we can use the Vecchia approximation to sample the integrand at a linear complexity with respect to n. Additionally, both the SOV algorithm from Genz (92) and the exponential-tilting method from Botev (2017) can be adapted to linear complexity.
}
\keyword{internal}
