\name{VeccTMVN-package}
\alias{VeccTMVN-package}
\alias{VeccTMVN}
\docType{package}
\title{
  Multivariate Normal Probabilities using Vecchia Approximation
}
\description{
  Under a different representation of the multivariate normal (MVN) probability, we can use the Vecchia approximation to sample the integrand at a linear complexity with respect to n. Additionally, both the SOV algorithm from Genz (92) and the exponential-tilting method from Botev (2017) can be adapted to linear complexity. However, this technique is only applicable to spatial covariance matrices.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jian Cao

Jian Cao <jcao2416@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
